requirements.txt
# Python packages for Windows + SQLite deployment
fastapi>=0.110,<0.116
uvicorn[standard]>=0.27,<0.31
python-multipart>=0.0.6,<0.0.10
jinja2>=3.1.2,<3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
librouteros>=3.2.0,<3.4
python-dotenv>=1.0.0,<1.1
pydantic>=2.7,<2.11
aiofiles==23.2.1
schedule>=1.2.0,<1.3
psutil>=5.9.8,<6.1
